requirements.txt
streamlit>=1.28.0
chromadb>=0.4.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-chroma>=0.1.0
langchain-ollama>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-experimental>=0.0.50
python-docx>=0.8.11
pandas>=1.5.0
numpy>=1.24.0
